requirements.txt
altair==5.0.1
attrs==23.1.0
backcall @ file:///C:/Program%20Files/RaySearch%20Laboratories/RayStation%2011A-SP2/ScriptClient/CPython38Dependencies/backcall-0.2.0-py2.py3-none-any.whl
backports.zoneinfo==0.2.1
blinker==1.6.2
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
colorama @ file:///C:/Program%20Files/RaySearch%20Laboratories/RayStation%2011A-SP2/ScriptClient/CPython38Dependencies/colorama-0.4.4-py2.py3-none-any.whl
contextlib2==21.6.0
contourpy==1.1.0
cycler==0.11.0
DateTime==5.2
decorator @ file:///C:/Program%20Files/RaySearch%20Laboratories/RayStation%2011A-SP2/ScriptClient/CPython38Dependencies/decorator-4.4.2-py2.py3-none-any.whl
et-xmlfile==1.0.1
fonttools==4.41.1
gitdb==4.0.10
GitPython==3.1.32
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.0
ipython @ file:///C:/Program%20Files/RaySearch%20Laboratories/RayStation%2011A-SP2/ScriptClient/CPython38Dependencies/ipython-7.19.0-py3-none-any.whl
ipython-genutils @ file:///C:/Program%20Files/RaySearch%20Laboratories/RayStation%2011A-SP2/ScriptClient/CPython38Dependencies/ipython_genutils-0.2.0-py2.py3-none-any.whl
jaraco.classes==3.3.1
jdcal @ file:///C:/Program%20Files/RaySearch%20Laboratories/RayStation%2011A-SP2/ScriptClient/CPython38Dependencies/jdcal-1.4.1-py2.py3-none-any.whl
jedi @ file:///C:/Program%20Files/RaySearch%20Laboratories/RayStation%2011A-SP2/ScriptClient/CPython38Dependencies/jedi-0.17.2-py2.py3-none-any.whl
Jinja2==3.1.2
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
keyring==24.3.0
kiwisolver==1.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.7.2
matplotlip==0.2
mdurl==0.1.2
more-itertools==10.2.0
numpy==1.23.1
openpyxl @ file:///C:/Program%20Files/RaySearch%20Laboratories/RayStation%2011A-SP2/ScriptClient/CPython38Dependencies/openpyxl-3.0.6-py2.py3-none-any.whl
packaging==23.1
pandas==2.0.3
parso==0.7.1
pickleshare==0.7.5
Pillow==9.5.0
pip==23.3.1
pkgutil_resolve_name==1.3.10
prompt-toolkit @ file:///C:/Program%20Files/RaySearch%20Laboratories/RayStation%2011A-SP2/ScriptClient/CPython38Dependencies/prompt_toolkit-3.0.16-py3-none-any.whl
protobuf==4.23.4
pyarrow==12.0.1
pycap==2.6.0
pycparser @ file:///C:/Program%20Files/RaySearch%20Laboratories/RayStation%2011A-SP2/ScriptClient/CPython38Dependencies/pycparser-2.20-py2.py3-none-any.whl
pydeck==0.8.0
Pygments==2.15.1
Pympler==1.0.1
pymssql==2.2.11
pyparsing==3.0.9
PyQt6==6.5.1
PyQt6-Qt6==6.5.1
PyQt6-sip==13.5.1
pyreadline==2.1
python-dateutil==2.8.2
python-tds==1.13.0
pythonnet @ file:///C:/Program%20Files/RaySearch%20Laboratories/RayStation%2011A-SP2/ScriptClient/CPython38Dependencies/pythonnet-2.5.1-cp38-cp38-win_amd64.whl
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
pywin32 @ file:///C:/Program%20Files/RaySearch%20Laboratories/RayStation%2011A-SP2/ScriptClient/CPython38Dependencies/pywin32-300-cp38-cp38-win_amd64.whl
pywin32-ctypes==0.2.2
PyYAML==6.0.1
referencing==0.30.0
requests==2.31.0
rich==13.4.2
rpds-py==0.9.2
scipy==1.10.1
semantic-version==2.10.0
setuptools==49.2.1
six==1.16.0
smmap==5.0.0
streamlit==1.25.0
tenacity==8.2.2
toml==0.10.2
toolz==0.12.0
tornado==6.3.2
traitlets @ file:///C:/Program%20Files/RaySearch%20Laboratories/RayStation%2011A-SP2/ScriptClient/CPython38Dependencies/traitlets-5.0.5-py3-none-any.whl
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==4.3.1
urllib3==2.0.4
validators==0.20.0
watchdog==3.0.0
wcwidth @ file:///C:/Program%20Files/RaySearch%20Laboratories/RayStation%2011A-SP2/ScriptClient/CPython38Dependencies/wcwidth-0.2.5-py2.py3-none-any.whl
zipp==3.16.2
zope.interface==6.0
